Moves business hours in header to its own separate component
