+++ components/Header/index.tsx
@@ -1,97 +1,14 @@
-import React, { useState, useEffect } from "react";
-import {
-  closeToClosing,
-  compressDays,
-  isOpen,
-  isWeekend,
-} from "utils/helpers/hours";
+import React, { useState } from "react";
 
 import styles from "./header.module.scss";
-import {
-  client,
-  GET_WEEKDAY_BUSINESS_HOURS,
-  GET_WEEKEND_BUSINESS_HOURS,
-} from "server/actions/Contentful";
-
-import { useQuery } from "@apollo/client";
-import { BusinessHoursResponse, CartItem, Item } from "utils/types";
+import { CartItem, Item } from "utils/types";
 import useSWR from "swr";
 import urls from "utils/urls";
-import DropdownLinks from "components/DropdownLinks";
+import StoreOpenStatus from "components/StoreOpenStatus";
+
 const Header: React.FC = () => {
   const [mobileOpen, setMobileOpen] = useState(false);
   //storeOpen, closingSoon, and storeStatus could probably all be refactored into one state.
-  const [storeOpen, setStoreOpen] = useState(false);
-  const [closingSoon, setClosingSoon] = useState(false);
-  const [storeStatus, setStoreStatus] = useState("");
-  const [bhOpen, setBHOpen] = useState(false);
-  const query = isWeekend()
-    ? GET_WEEKEND_BUSINESS_HOURS
-    : GET_WEEKDAY_BUSINESS_HOURS;
-  const otherQuery = isWeekend()
-    ? GET_WEEKDAY_BUSINESS_HOURS
-    : GET_WEEKEND_BUSINESS_HOURS; // Get the remaining query
-
-  const {
-    loading: isOpenLoading,
-    data: isOpenData,
-    error: isOpenError,
-  } = useQuery<BusinessHoursResponse>(query, {
-    client: client,
-    pollInterval: 3600000,
-  });
-  const {
-    loading: otherLoading,
-    data: otherData,
-    error: otherError,
-  } = useQuery<BusinessHoursResponse>(otherQuery, {
-    client: client,
-    pollInterval: 3600000, //Poll every hour
-  });
-
-  useEffect(() => {
-    if (isOpenData) {
-      if (isOpen(isOpenData.businessHoursCollection.items[0])) {
-        setStoreOpen(true);
-        if (closeToClosing(isOpenData.businessHoursCollection.items[0].hours)) {
-          setClosingSoon(true);
-        }
-      } else {
-        setStoreOpen(false);
-      }
-      if (storeOpen && closingSoon) {
-        setStoreStatus("Closing soon.");
-      } else if (storeOpen && !closingSoon) {
-        setStoreStatus("We are open!");
-      } else {
-        setStoreStatus("We are closed.");
-      }
-    }
-    const interval = setInterval(() => {
-      if (isOpenData) {
-        if (isOpen(isOpenData.businessHoursCollection.items[0])) {
-          setStoreOpen(true);
-          if (
-            closeToClosing(isOpenData.businessHoursCollection.items[0].hours)
-          ) {
-            setClosingSoon(true);
-          }
-        } else {
-          setStoreOpen(false);
-        }
-
-        if (storeOpen && closingSoon) {
-          setStoreStatus("Closing soon.");
-        } else if (storeOpen && !closingSoon) {
-          setStoreStatus("We are open!");
-        } else {
-          setStoreStatus("We are closed.");
-        }
-      }
-    }, 60000);
-
-    return () => clearInterval(interval);
-  }, [isOpenData, storeOpen, closingSoon]);
 
   const fetcher = (url: string): Promise<unknown> =>
     fetch(url).then(r => r.json());
@@ -104,59 +21,10 @@ const Header: React.FC = () => {
   };
   return (
     <div>
-      <DropdownLinks category="trevor" />
       <header className={styles.headerParent}>
         <div className={styles.upperHeader}>
           <div className={styles.upperHeaderLeft}>
-            {isOpenData && !isOpenError && !isOpenLoading && (
-              <div className={styles.hours}>
-                <div
-                  data-is-open={storeOpen}
-                  data-closing-soon={closingSoon}
-                ></div>
-                <span
-                  role="button"
-                  tabIndex={0}
-                  onClick={() =>
-                    window.innerWidth < 1100
-                      ? setBHOpen(!bhOpen)
-                      : setBHOpen(!bhOpen)
-                  }
-                  onKeyDown={(e: React.KeyboardEvent) => {
-                    if (e.key == "enter") {
-                      setBHOpen(!bhOpen);
-                    }
-                  }}
-                >
-                  {storeStatus}
-                </span>
-                <div
-                  className={`${styles.businessHoursSub} ${
-                    bhOpen ? styles.businessHoursSubActive : ""
-                  }`}
-                >
-                  <div className={styles.navButtonSubTriangle}></div>
-                  {otherData && !otherLoading && !otherError && (
-                    <p>
-                      {compressDays(
-                        otherData.businessHoursCollection.items[0].daysOpen
-                      )}
-                      {otherData.businessHoursCollection.items[0].hours.join(
-                        ", "
-                      )}
-                    </p>
-                  )}
-                  <p>
-                    {compressDays(
-                      isOpenData.businessHoursCollection.items[0].daysOpen
-                    )}
-                    {isOpenData.businessHoursCollection.items[0].hours.join(
-                      ", "
-                    )}
-                  </p>
-                </div>
-              </div>
-            )}
+            <StoreOpenStatus />
           </div>
           <div className={styles.upperHeaderRight}>
             <a href="/">Find Us</a>
